+++ packages/xstate-vue/src/index.ts
@@ -2,3 +2,4 @@ export { useMachine } from './useMachine';
 export { useService } from './useService';
 export { useActor } from './useActor';
 export { useInterpret } from './useInterpret';
+export { useSelector } from './useSelector';